get_ai_input.py
#!/usr/bin/env python3
"""
Direct AI consultation for 3-Tier Evolutionary Trading System Architecture
"""

import google.generativeai as genai
import json
from pathlib import Path

# Configure Gemini
API_KEY = "AIzaSyAZFQzr1YcGBgmt2CVnECoM_-JALCg5uZw"
genai.configure(api_key=API_KEY)
model = genai.GenerativeModel('models/gemini-1.5-flash')

def get_gemini_architecture_input():
    """Get Gemini's input on the 3-Tier Evolutionary Trading System architecture"""
    
    prompt = """
    I'm building a sophisticated 3-Tier Evolutionary Trading System for QuantConnect with the following specifications:

    **PERFORMANCE TARGETS:**
    - 25% CAGR (Compound Annual Growth Rate)
    - 1.0+ Sharpe Ratio (Risk-adjusted return) 
    - <15% Maximum Drawdown (Risk management)

    **TECHNICAL REQUIREMENTS:**
    - Strategy Generation Rate: 100+ strategies per hour
    - 15-year historical validation (2009-2024) in <30 minutes
    - API Response Time: <100ms average
    - Memory Usage: <16GB peak
    - System Initialization: <1 minute

    **QUANTCONNECT INTEGRATION:**
    - User ID: 357130
    - Token: 62d0a329b3c854f1f61d29114eb02a7c724b361878a85d7953d0ba0e2b053912
    - 400TB data library access
    - Multi-asset support (Equities, Options, Futures, CFD, Forex, Crypto)
    - Real-time and historical data feeds

    **CURRENT ARCHITECTURE:**
    - **TIER 1**: Core Execution Engine (main.py → controller.py → config.py)
    - **TIER 2**: Strategy Generation & Testing with 1000+ existing strategies  
    - **TIER 3**: Advanced Evolution Systems with multi-agent optimization

    **MULTI-AGENT SYSTEM (6 Agents):**
    1. Supervisor Agent - Overall coordination
    2. Trend Following Agent - Trend-based strategies
    3. Mean Reversion Agent - Mean reversion strategies  
    4. Momentum Agent - Momentum-based trading
    5. Arbitrage Agent - Arbitrage opportunities
    6. Market Neutral Agent - Market-neutral strategies

    **TECHNOLOGY STACK:**
    - Python 3.11+ with LEAN Algorithm Framework
    - DEAP 1.4.3 (Distributed Evolutionary Algorithms)
    - NumPy 2.1.4+, Pandas 2.1.4
    - QuantConnect Cloud Platform
    - Real-time monitoring with 1-second updates

    **QUESTIONS FOR YOU:**

    1. **Architecture Optimization**: What improvements would you suggest to the 3-tier architecture to achieve the aggressive performance targets (25% CAGR, 1.0+ Sharpe)?

    2. **Strategy Generation**: How can we optimize the strategy generation process to reliably produce 100+ strategies per hour while maintaining quality?

    3. **Multi-Agent Coordination**: What coordination mechanisms between the 6 agents would be most effective for collaborative strategy discovery?

    4. **Risk Management**: What additional risk management layers should be integrated into each tier to ensure the <15% drawdown target?

    5. **Performance Optimization**: What specific algorithmic optimizations would you recommend for meeting the <30 minute backtesting requirement over 15 years of data?

    6. **Evolution Algorithm**: How should we structure the genetic algorithm parameters (population size, mutation rates, selection pressure) for optimal strategy evolution?

    7. **Real-time Monitoring**: What key metrics and thresholds should the 1-second monitoring system track for early detection of performance degradation?

    8. **Scalability**: How can we design the system to scale beyond the initial requirements while maintaining performance targets?

    Please provide detailed, technical recommendations for each area. Focus on practical implementation strategies that can achieve these ambitious targets.
    """
    
    try:
        print("🧠 Consulting Gemini on architecture design...")
        response = model.generate_content(prompt)
        return response.text
    except Exception as e:
        return f"Error getting Gemini input: {str(e)}"

def save_ai_input(content: str, filename: str):
    """Save AI input to file"""
    docs_dir = Path("docs")
    docs_dir.mkdir(exist_ok=True)
    
    with open(docs_dir / filename, 'w') as f:
        f.write(content)
    print(f"✅ Saved AI input to docs/{filename}")

if __name__ == "__main__":
    print("=" * 80)
    print("🤖 GETTING AI INPUT FOR 3-TIER EVOLUTIONARY TRADING SYSTEM")
    print("=" * 80)
    
    # Get Gemini's input
    gemini_input = get_gemini_architecture_input()
    save_ai_input(gemini_input, "gemini_architecture_input.md")
    
    print("\n" + "=" * 80)
    print("📋 GEMINI'S ARCHITECTURE RECOMMENDATIONS:")
    print("=" * 80)
    print(gemini_input)
    
    print("\n" + "=" * 80)
    print("✅ AI consultation complete! Check docs/gemini_architecture_input.md for full details.")
    print("=" * 80)