requirements.txt
# 3-Tier Evolutionary Trading System Requirements

# Core scientific computing
numpy>=2.1.4
pandas>=2.1.4
scipy>=1.11.0

# Evolutionary algorithms
deap>=1.4.3

# Async and networking
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
websockets>=12.0

# QuantConnect and trading
quantconnect>=1.0.0
lean>=1.0.0

# Machine learning (for strategy optimization)
scikit-learn>=1.3.0
xgboost>=2.0.0
optuna>=3.4.0

# Data analysis and visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Configuration and utilities
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7

# Testing and quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Monitoring and logging
psutil>=5.9.0
structlog>=23.2.0

# Security
cryptography>=41.0.0
bcrypt>=4.1.0

# Database (for strategy storage)
sqlalchemy>=2.0.0
alembic>=1.13.0
sqlite3  # Built-in

# HTTP clients
requests>=2.31.0
httpx>=0.25.0

# Time series analysis
statsmodels>=0.14.0
arch>=6.2.0  # GARCH models

# Risk management
pyfolio>=0.9.2
empyrical>=0.5.5

# Data providers (optional)
yfinance>=0.2.0
alpha-vantage>=2.3.1
polygon-api-client>=1.12.0

# Development tools
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Research and web crawling (for open-deep-research integration)
firecrawl-py>=0.0.20
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
textblob>=0.17.1
spacy>=3.7.0
transformers>=4.36.0
openai>=1.0.0
anthropic>=0.7.0

# Intelligent Memory System Dependencies
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
networkx>=3.0